app/requirements.txt
flask==0.11.1
flask-restplus==0.9.2
requests==2.3.0
httplib2==0.9.2
elasticsearch==5.0.1
PyJWT==1.4.2
python-dotenv==0.6.0
